feat: segregate sendnn XML ingestion into standalone script
